+++ src/app/(auth)/sign-in/page.tsx
@@ -1,7 +1,0 @@
-import { SignIn } from '@/components/sign-in';
-
-const page = () => {
-  return <SignIn />;
-};
-
-export default page;