+++ src/app/dashboard/page.tsx
@@ -1,0 +1,15 @@
+'use client';
+import { Button } from '@/components/ui/button';
+import { useStoreAdminLogout } from '@/services/base/store-admin/useStoreAdminLogout';
+
+const DashboardPage = () => {
+  const logoutMutation = useStoreAdminLogout();
+
+  return (
+    <div>
+      <Button onClick={() => logoutMutation.mutate()}>Logout</Button>
+    </div>
+  );
+};
+
+export default DashboardPage;