+++ src/schemas/storeAdminSignIn.ts
@@ -1,7 +1,0 @@
-import { z } from 'zod';
-import { mobileNumberValidation, passwordValidation } from './storeAdminSignUp';
-
-export const storeAdminSignInSchema = z.object({
-  mobileNumber: mobileNumberValidation,
-  password: passwordValidation,
-});